Add VPN gateway and customer gateway to VPC config file
This commit adds a VPN gateway and a customer gateway to the VPC configuration file, along with a VPN connection and a route to a remote private subnet. The configuration file also propagates routes from the VPN gateway to the VPC route table for secure communication with the remote subnet.

On the remote side, a pfsense router was configured using the web console and pfsense internal IPsec capabilities. After creating the tunnel, the AWS security groups had to be modified through the web console to allow traffic to the private subnets. In the future, the security group configuration will be moved to Terraform, and the scripts will be added to this repository.

Changes Made:
- Added VPN gateway and customer gateway to VPC configuration file
- Created VPN connection and route to remote private subnet
- Configured pfsense router using web console and internal IPsec capabilities
- Modified AWS security groups to allow traffic to private subnets

Future Improvements:
- Move security group configuration to Terraform and add scripts to repository